Fixed home screen post issue

+++ lib/model/videoInfoModel.dart
@@ -18,25 +18,24 @@ class VideoInfo {
   int comments;
   String videoId;
 
-  VideoInfo(
-      {
-        this.uploaderUid,
-        this.videoUrl,
-        this.thumbUrl,
-        this.coverUrl,
-        this.shareUrl,
-        this.aspectRatio,
-        this.uploadedAt,
-        this.videoName,
-        this.videoHashtag,
-        this.category,
-        this.views,
-        this.likes,
-        this.shares,
-        this.rating,
-        this.comments,
-        this.videoId,
-      });
+  VideoInfo({
+    this.uploaderUid,
+    this.videoUrl,
+    this.thumbUrl,
+    this.coverUrl,
+    this.shareUrl,
+    this.aspectRatio,
+    this.uploadedAt,
+    this.videoName,
+    this.videoHashtag,
+    this.category,
+    this.views,
+    this.likes,
+    this.shares,
+    this.rating,
+    this.comments,
+    this.videoId,
+  });
 
   static fromDocument(QueryDocumentSnapshot ds) {
     return VideoInfo(
@@ -48,7 +47,7 @@ class VideoInfo {
       rating: ds.data()['rating'],
       likes: ds.data()['likes'],
       comments: ds.data()['comments'],
-      aspectRatio: ds.data()['aspectRatio'],
+      aspectRatio: ds.data()['aspectRatio'].toDouble(),
       videoHashtag: ds.data()['videoHashtag'],
       videoName: ds.data()['videoName'],
       uploadedAt: ds.data()['uploadedAt'],
